:recycle: Separate reading animation from state
