+++ src/friendo/element/fire.js
@@ -80,7 +80,9 @@ export default class Fire extends Element {
 
       // handle glasses
       // doesn't need glasses to see anymore after 9
-      if (friendo.state.glasses && friendo.getStatStage(STATS.SIGHT) < 3) threeLens(g, x, y + 3)
+      if (friendo.state.anim.glasses && friendo.getStatStage(STATS.SIGHT) < 4) {
+        threeLens(g, x, y + 3)
+      }
     } else if (friendo.getStatStage(STATS.SIGHT) > 1) {
       // lvl 4 and up, 2 eyes
       // eyes must be moved down if a fire element
@@ -88,13 +90,13 @@ export default class Fire extends Element {
       this.drawEye(g, x + 6, y + 4, doBlink, friendo.state.anim.isSmiling)
 
       // handle glasses
-      if (friendo.state.glasses) twoLens(g, x, y + 3)
+      if (friendo.state.anim.glasses) twoLens(g, x, y + 3)
     } else {
       // default = 1 eye
       this.drawEye(g, x, y, doBlink, friendo.state.anim.isSmiling)
 
       // handle glasses
-      if (friendo.state.glasses) oneLens(g, x, y)
+      if (friendo.state.anim.glasses) oneLens(g, x, y)
     }
 
     drawHookMarker(g, x, y)
